chore: moved helper functions to seperate utils file
